third_party/WebKit/LayoutTests/http/tests/inspector/bindings/dynamic-bindings-frame-attach-detach.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="./bindings-test.js"></script>

<script>

async function test() {
  TestRunner.markStep('dumpInitialWorkspace');
  var snapshot = BindingsTestRunner.dumpWorkspace();

  TestRunner.markStep('attachFrame');
  await BindingsTestRunner.attachFrame('frame', './resources/dynamic-frame.html', '_test_attachFrame.js');
  snapshot = BindingsTestRunner.dumpWorkspace(snapshot);

  TestRunner.markStep('detachFrame');
  await BindingsTestRunner.detachFrame('frame', '_test_detachFrame.js');
  snapshot = BindingsTestRunner.dumpWorkspace(snapshot);

  TestRunner.completeTest();
}

</script>

</head>
<body onload="runTest()">
<p>
Verify that UISourceCodes are added and removed as iframe with dynamic script
and stylesheet is added and removed.
</p>
</body>
</html>
