third_party/WebKit/LayoutTests/http/tests/inspector/bindings/dynamic-bindings-frame-attach-detach-expected.txt

Verify that UISourceCodes are added and removed as iframe with dynamic script and stylesheet is added and removed.


Running: dumpInitialWorkspace
Removed: 0 uiSourceCodes
Workspace: 7 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] dynamic-bindings-frame-attach-detach.html
    debugger:///VM[XXX] dynamic-bindings-frame-attach-detach.html
    debugger:///VM[XXX] inspector-test.js
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/dynamic-bindings-frame-attach-detach.html
    http://127.0.0.1:8000/inspector/inspector-test.js

Running: attachFrame
Removed: 0 uiSourceCodes
Workspace: 14 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] dynamic-bindings-frame-attach-detach.html
    debugger:///VM[XXX] dynamic-bindings-frame-attach-detach.html
[+] debugger:///VM[XXX] dynamic-frame.html
[+] debugger:///VM[XXX] dynamic-script.js
    debugger:///VM[XXX] inspector-test.js
[+] debugger:///VM[XXX] _test_attachFrame.js
[+] dynamic-script.js
[+] dynamic-style.css
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/dynamic-bindings-frame-attach-detach.html
[+] http://127.0.0.1:8000/inspector/bindings/resources/dynamic-frame.html
    http://127.0.0.1:8000/inspector/inspector-test.js
[+] _test_attachFrame.js

Running: detachFrame
Removed: 2 uiSourceCodes
[-] dynamic-style.css
[-] http://127.0.0.1:8000/inspector/bindings/resources/dynamic-frame.html
Workspace: 14 uiSourceCodes.
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] dynamic-bindings-frame-attach-detach.html
    debugger:///VM[XXX] dynamic-bindings-frame-attach-detach.html
    debugger:///VM[XXX] dynamic-frame.html
    debugger:///VM[XXX] dynamic-script.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] _test_attachFrame.js
[+] debugger:///VM[XXX] _test_detachFrame.js
    dynamic-script.js
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/dynamic-bindings-frame-attach-detach.html
    http://127.0.0.1:8000/inspector/inspector-test.js
    _test_attachFrame.js
[+] _test_detachFrame.js
